move components from form page dir to components dir
